clients/mcp-client/pyproject.toml
[project]
name = "mcp-client"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "anthropic>=0.83.0",
    "mcp[cli]>=1.26.0",
    "python-dotenv>=1.2.1",
]
